[PAIG-2210] Removed dummy llama transformer code
